seperate user management factors from normal factors
