+++ src/mfa/interfaces/factor.interface.ts
@@ -13,16 +13,6 @@ export interface Factor {
   totp?: Totp;
 }
 
-export interface UserManagementFactor {
-  object: 'authentication_factor';
-  id: string;
-  createdAt: string;
-  updatedAt: string;
-  type: 'totp';
-  totp: Totp;
-  userId: string;
-}
-
 export interface FactorResponse {
   object: 'authentication_factor';
   id: string;
@@ -32,14 +22,4 @@ export interface FactorResponse {
   sms?: SmsResponse;
   totp?: TotpResponse;
   user_id?: string;
-}
-
-export interface UserManagementFactorResponse {
-  object: 'authentication_factor';
-  id: string;
-  created_at: string;
-  updated_at: string;
-  type: 'totp';
-  totp: TotpResponse;
-  user_id: string;
 }
